requirements.txt
connexion[swagger-ui]
PyYAML
httpx
APScheduler
pymongo 